update command, query, update dependency

+++ lib/SQLAPI/include/asaAPI.h
@@ -106,7 +106,7 @@ public:
 	virtual void InitializeClient(const SAOptions *pOptions);
 	virtual void UnInitializeClient(const SAPI *pSAPI, const SAOptions* pOptions);
 
-	virtual long GetClientVersion() const;
+	virtual int GetClientVersion() const;
 
 	virtual ISAConnection *NewConnection(SAConnection *pConnection);
