+++ lib/SQLAPI/include/oraAPI.h
@@ -675,14 +675,14 @@ public:
 	virtual void InitializeClient(const SAOptions *pOptions);
 	virtual void UnInitializeClient(const SAPI *pSAPI, const SAOptions* pOptions);
 
-	virtual long GetClientVersion() const;
+	virtual int GetClientVersion() const;
 
 	virtual ISAConnection *NewConnection(SAConnection *pConnection);
 
 protected:
 	void  *m_hLibrary;
 	SAMutex m_loaderMutex;
-	long m_nORA8DLLVersionLoaded;
+	int m_nORA8DLLVersionLoaded;
 #ifdef SA_UNICODE
 	bool m_bUseUCS2;
 #endif
